ecom_stream_processing/spark_streaming/requirements.txt
pyspark==3.5.5
PyYAML==6.0.2
PyYAML==6.0.2
